requirements.txt
asgiref==3.8.1
bcrypt==4.2.1
Django==5.1.3
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
httpcore==1.0.7
httpx==0.27.2
# Change path to your MATLAB installation
matlabengine @ file:///Applications/MATLAB_R2024b.app/extern/engines/python
psycopg2-binary==2.9.10
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
supabase==2.10.0
supafunc==0.7.0

# Machine Learning dependencies
numpy>=1.24.0
scikit-learn>=1.2.0
joblib>=1.3.0
pandas>=1.5.0
xgboost>=1.7.0
matplotlib>=3.7.0